Fixed some experience stuff not being longs, added a formula class, and changed the entity box label
Former-commit-id: a650d64386279625796e0cb0706cfd640eff11cf

+++ Intersect Client/Classes/Entities/Player.cs
@@ -28,8 +28,8 @@ namespace Intersect_Client.Classes.Entities
         public int TargetType;
 
         public int Class = -1;
-        public int Experience = 0;
-        public int ExperienceToNextLevel = 0;
+        public long Experience = 0;
+        public long ExperienceToNextLevel = 0;
         public List<FriendInstance> Friends = new List<FriendInstance>();
         public HotbarInstance[] Hotbar = new HotbarInstance[Options.MaxHotbar];
 
@@ -910,7 +910,7 @@ namespace Intersect_Client.Classes.Entities
         }
 
         //Forumlas
-        public int GetNextLevelExperience()
+        public long GetNextLevelExperience()
         {
             return ExperienceToNextLevel;
         }
